requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.7.0
django-filter==25.1
django-model-utils==5.0.0
djangorestframework==3.15.2
dnspython==2.7.0
filelock==3.18.0
idna==3.10
kombu==5.5.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
py3-validate-email==1.0.5.post2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
